Complete 01-css-methods
Firstly, was created a file - style.css. Then linked style.css to the index.html

Added <style></style> with properties and values for the internal method and also style as an attribute for the inline method in our <button>

Also was added type selector with properties and values to the style.css - for the external method.